+++ server-project/CustomerAccount.WebAPI/DTOs/AccountInfoDTO.cs
@@ -4,6 +4,6 @@
     {
         public string FirstName { get; set; }
         public string LastName { get; set; }
-        public double Balance { get; set; }
+        public int Balance { get; set; }
     }
 }
